Converted math parser to .net core Char.ToString does not take CultureInfo now
