+++ MathParser/MathParser.sln
@@ -1,31 +1,26 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.26403.3
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathParser", "MathParser.csproj", "{E17D08FF-ADF8-4F6E-B6D9-ED343E76E3A8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathParser", "MathParser\MathParser.csproj", "{73FFDCCC-F886-425F-9776-5E94665996BA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathParserTest", "UnitTest\MathParserTest.csproj", "{4BC67DFD-E614-4E23-91D6-25F38D575BA5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathParserTests", "MathParserTests\MathParserTests.csproj", "{CF8CA78F-DE6E-43EC-9DE6-915D5EB927A6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
-		ReleaseNUnit|Any CPU = ReleaseNUnit|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E17D08FF-ADF8-4F6E-B6D9-ED343E76E3A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E17D08FF-ADF8-4F6E-B6D9-ED343E76E3A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E17D08FF-ADF8-4F6E-B6D9-ED343E76E3A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E17D08FF-ADF8-4F6E-B6D9-ED343E76E3A8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E17D08FF-ADF8-4F6E-B6D9-ED343E76E3A8}.ReleaseNUnit|Any CPU.ActiveCfg = Release|Any CPU
-		{E17D08FF-ADF8-4F6E-B6D9-ED343E76E3A8}.ReleaseNUnit|Any CPU.Build.0 = Release|Any CPU
-		{4BC67DFD-E614-4E23-91D6-25F38D575BA5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4BC67DFD-E614-4E23-91D6-25F38D575BA5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4BC67DFD-E614-4E23-91D6-25F38D575BA5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4BC67DFD-E614-4E23-91D6-25F38D575BA5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4BC67DFD-E614-4E23-91D6-25F38D575BA5}.ReleaseNUnit|Any CPU.ActiveCfg = ReleaseNUnit|Any CPU
-		{4BC67DFD-E614-4E23-91D6-25F38D575BA5}.ReleaseNUnit|Any CPU.Build.0 = ReleaseNUnit|Any CPU
+		{73FFDCCC-F886-425F-9776-5E94665996BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{73FFDCCC-F886-425F-9776-5E94665996BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{73FFDCCC-F886-425F-9776-5E94665996BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{73FFDCCC-F886-425F-9776-5E94665996BA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CF8CA78F-DE6E-43EC-9DE6-915D5EB927A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CF8CA78F-DE6E-43EC-9DE6-915D5EB927A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CF8CA78F-DE6E-43EC-9DE6-915D5EB927A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CF8CA78F-DE6E-43EC-9DE6-915D5EB927A6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
